konek to fe, buy sell portofolio

+++ src/main/java/com/project/crypto/service/CryptoUserService.java
@@ -1,7 +1,9 @@
 package com.project.crypto.service;
 
 import java.util.List;
+import java.util.Optional;
 
+import com.project.crypto.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -18,8 +20,12 @@ public class CryptoUserService {
 	
 	@Autowired
 	private CryptoUserRepository cryptoUserRepository;
+
+	@Autowired
+	UserRepository userRepository;
 	
-	public List<CryptoUser> getList(Integer userId) {
+	public List<CryptoUser> getList(String username) {
+		Integer userId = userRepository.findByUsername(username).getId();
 		try {
 			logCryp.logCrypBe.info("get list crypto user "+userId);
 			return cryptoUserRepository.getList(userId);
@@ -28,11 +34,30 @@ public class CryptoUserService {
 		}
 		return null;
 	}
+
+	public String checkNum(Integer id, Long jumlah){
+		Optional<CryptoUser> cryptoUser = cryptoUserRepository.findById(id);
+		if (cryptoUser.isPresent()){
+			CryptoUser crypto = cryptoUser.get();
+			if(jumlah <= crypto.getJumlah()) return "berhasil";
+		}
+		return "gagal";
+	}
 	
-	public void deleteTxn(int id) {
+	public void deleteTxn(int id, Long jumlah) {
 		try {
 			logCryp.logCrypBe.info("penghapusan txn di postofolio id="+id);
-			cryptoUserRepository.deleteTxn(id);
+			Optional<CryptoUser> cryptoUser = cryptoUserRepository.findById(id);
+			if (cryptoUser.isPresent()){
+				CryptoUser crypto = cryptoUser.get();
+				if (crypto.getJumlah().equals(jumlah)){
+					cryptoUserRepository.deleteTxn(id);
+				}else{
+					crypto.setJumlah(crypto.getJumlah() - jumlah);
+					cryptoUserRepository.save(crypto);
+				}
+			}
+//			cryptoUserRepository.deleteTxn(id);
 		} catch (Exception e) {
 			logCryp.logCrypBe.error("delete txn di portofolio gagal: "+e);
 		}
